+++ CHANGELOG.md
@@ -32,6 +32,7 @@ for further information about branching and tagging conventions.
 ### Fixed
 - Using account roles instead of account types for limits requests
 - Using `request.details.updateLimits` instead of `request.details.limitsUpdate`
+- `localizeIssuanceRequestState` filter
 
 ## [1.2.0-rc.9] - 2019-02-25
 ### Change